feat: 기능 대부분 구현

+++ src/main/kotlin/kr/hanghae/deploy/dto/user/ChargeBalanceServiceRequest.kt
@@ -1,11 +1,13 @@
 package kr.hanghae.deploy.dto.user
 
+import java.math.BigDecimal
+
 data class ChargeBalanceServiceRequest(
-    val balance: Long,
+    val balance: BigDecimal,
     val uuid: String,
 ) {
     companion object {
-        fun toService(balance: Long, uuid: String): ChargeBalanceServiceRequest {
+        fun toService(balance: BigDecimal, uuid: String): ChargeBalanceServiceRequest {
             return ChargeBalanceServiceRequest(
                 balance = balance,
                 uuid = uuid,
