+++ src/main/kotlin/kr/hanghae/deploy/dto/user/request/ChargeBalanceRequest.kt
@@ -1,5 +1,7 @@
 package kr.hanghae.deploy.dto.user.request
 
+import java.math.BigDecimal
+
 data class ChargeBalanceRequest(
-    val balance: Long,
+    val balance: BigDecimal,
 )
